Move vote_extensions to its own file
